remove bufpoll subdir and put inside render
